Patch some Windows C code that breaks under gcc/clang. (#15326)
Fully supporting gcc and clang on Windows is proving to be tricky. This
at least gets the builds a bit further along.

ci-extra: build_test_all_windows

+++ runtime/src/iree/base/status.c
@@ -628,8 +628,8 @@ static bool iree_symbol_resolver_format_frame(iree_symbol_resolver_t* resolver,
   if (resolver->SymGetLineFromAddr64(process, (DWORD64)address, &displacement32,
                                      &line)) {
     *buffer_length = iree_string_buffer_append_format(
-        buffer_capacity, buffer, *buffer_length, " (%s:%u)", line.FileName,
-        line.LineNumber);
+        buffer_capacity, buffer, *buffer_length, " (%s:%" PRIu32 ")",
+        line.FileName, (uint32_t)line.LineNumber);
   }
   iree_symbol_resolver_unlock(resolver);
